fix path in files.lst and disable tar checking
Fix the path in the files.lst file from containing the groupDir to only contain
the pnfsid.
Deactivate the checking of the tar files doing a tar tf and comparing files count.
This caused problems because the tars are visible for a while as 0-length file and
tar tf fails which caused the archives to be deleted.