simplify file collection algorithm
remote verify step as it does not really prevent problems with deleted flags.
Instead, an additional check now checks if the whole collection succeeded and
stops the process if the script was hit by a deleted flag.

The detect deleted method is now included in the collectFiles method. This way
we avoid double requests to chimera.