consolidate packing scripts
the two packing scripts share most logic. The only difference is
in obtaining the group directories and how to extract the tags from
this directory.

This patch merges the two scripts and has the differences extracted
in two scripts containing the mode specific functions. The mode can
be chosen by passing by-group.sh or by-dir.sh as first parameter to
the master script pack-files.sh.