split pack-files.py into twi script
instead of only one script that handles both, filling the db
with the files metadata and creating the archives, the script
is split up into the scripts fillmetadata.py and pack-files.py.
fillmetadata.py runs in an infinite loop checking the db for
new files and pack-files.py runs an infinite loop collecting
"known" files and packing them.