add state entry to trace files
file entries in the mongo db can now have different state.  The
fillmetadata script will set the state to 'new' The pack-files script
will set the state to 'added' as soon as the file is in any container.
If the container is closed and stored the state will be set to
'archived'. If the container is deleted, the state will be reset to
'new'. The writebfid script will set the state to verified as soon as it
has entered the bfid as archiveUrl into the file entry in the db.