use exception for cleanup
if the script is interrupted it sets a flag. This flag is picked
up at several places to ensure a graceful shutdown. While it works, it
still creates a lot of duplicate code.
This patch introduces a new exception class that allows the script to
handle the cleaning up at exactly one place.