add field lock and scriptId
if the database connection fails during adding of files, some files
will be left in state added and not be picked up again. This needs
manual intervention to fix.

This patch adds an additional field to the database 'lock' that holds
the scriptId of the packing script. If the packing script finds files
that were locked by itself it will reset these to state 'new' and
thus make manual intervention unnecessary.