fix problem with expected filenumber
If the pack-files script switches into old-file mode in the file
number and size collection phase before actually adding the files to the
container, then, if more files arrive between that and the actual
packing, it may happen that the container that actually contains the old
files will have the regular archiveSize, but the following archive -
still being packed in old-file-mode - may actually contain no old files
and be smaller than the target archiveSize.

This patch ensures the packager is left when all files from the collection
phase have been archived.