manually merge old-files container fix
if old files were detected the old-file mode is activated.
There might be a discrepancy between the files that were checked
during the initial state and when the actual packing happens.
This could cause containers to be created that are smaller than
the target size, but don't contain any small files.
This patch takes care that only the files that were found in the
initial state are taken into account for packing